+++ src/main/java/mod/gottsch/forge/gottschcore/block/StructureMobSetBlock.java
@@ -1,0 +1,45 @@
+/*
+ * This file is part of Treasure2.
+ * Copyright (c) 2025 Mark Gottschling (gottsch)
+ *
+ * Treasure2 is free software: you can redistribute it and/or modify
+ * it under the terms of the Open Software Licence 3.0.
+ *
+ * Treasure2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Open Software Licence 3.0 for more details.
+ *
+ * You should have received a copy of the Open Software Licence
+ * along with Treasure2. If not, see <https://www.tldrlegal.com/license/open-software-licence-3-0>.
+ */
+package mod.gottsch.forge.gottschcore.block;
+
+import mod.gottsch.forge.gottschcore.block.entity.StructureMobSetBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.block.BaseEntityBlock;
+import net.minecraft.world.level.block.RenderShape;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ * This block is meant for Jigsaw Structures only. It's purpose is to act as a marker and
+ * store data for a StructureProcessor to utilize.
+ * @author by Mark Gottschling on 8/14/2025
+ */
+public abstract class StructureMobSetBlock extends BaseEntityBlock {
+
+    public StructureMobSetBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public abstract BlockEntity newBlockEntity(BlockPos pos, BlockState state);
+
+    @Override
+    public RenderShape getRenderShape(BlockState state) {
+        return RenderShape.MODEL;
+    }
+
+}